feat: add optional error callback when something goes wrong
With google or other providers you can get following error when you are browsing in private modus:
````error: "idpiframe_initialization_failed", details: "Cookies are not enabled in current environment."```

In this case I want to inform the user instead of having a silence error. My proposal is to add an erroCallback, where you can add an alert or something else to avoid that the user is not informed.